Stop using windeployqt
Our custom compiled Qt,version 5.9.7, has a defected  version
of windeployqt. Upon running it complains with the following
message: "unable to find platform plugins". Qt's bugs page
does have this mentioned as a bug:
https://bugreports.qt.io/browse/QTBUG-42072
Although it's status is marked as closed/merged but somehow
it's still not working for us.

For now, resorting to typical copy paste magic to get all the
requried plugins